Add a Lua library to communicate with DWS.
The Lua library expects to use kubectl to communicate with the K8s environment,
to manipulate the DWS API.  Include a test for the library.

Add a tools/ directory with a tool that creates a KIND (kubernetes-in-docker)
environment for testing this library with DWS.  Add a README to explain how
to use that environment.

Add makefile targets to test the library against a live system or to test it
with low-level functions stubbed.  Add a TAP output module for this local
testing.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>