Support a ServiceAccount token
Use the ServiceAccount token on the kubectl commandline when running inside
the slurm pod.

Move the replacements for the slurm log functions to the test file.

The Workflow CRD name must be fully-qualified to insulate us from other
services that use that name for their own CRDs.

Adjust the return values of a few DWS methods to avoid mixing strings with
tables.

Tighten the parsing for status values.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>