Verify or add a kubectl cache dir
Kubectl wants to place its discovery cache and http cache into $HOME/.kube,
however the PE slurmctld container may not have provided a home dir for the
user.  This detects the absence of a home dir and instead creates a dir in /tmp and sets a HOME env variable for the kubectl command to follow.

This can reduce the runtime of a typical kubectl command, run in the PE
slurmctld pod, from 25 seconds down to 0.4 seconds.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>