Adjust k8s env
Update the k8s version to 1.27.2 to match the version being used by DWS in the
Rabbit project.

Drop the cray.wlm.manager node label; DWS v0.0.11 doesn't use it.

Upgrade cert-manager to 1.11.1 to match the version being used by DWS in the
Rabbit project.

DWS v0.0.11 has a separate pod for the webhook, so wait for that to start.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>