Handle DWS fatal errors
Set Flags=TeardownFailure in burst_buffer.conf.  This tells Slurm to go to
Teardown if there are errors during stage_in or stage_out.

In the burst_buffer.lua plugin, adjust slurm_bb_job_teardown() to check
the workflow for any fatal errors recorded in the drivers array.  If any
are found, then we know we were called in a fatal error situation and we
not only delete the workflow, per usual, but we also call the Slurm `scancel`
command to cancel the Slurm job.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>